requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.4.1
Babel==2.9.1
billiard==3.6.4.0
braintree==3.59.0
cached-property==1.5.2
celery==4.4.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-localflavor==3.1
django-parler==2.2
django-rosetta==0.9.7
flower==0.9.3
humanize==3.11.0
idna==3.2
importlib-metadata==4.8.1
kombu==4.6.11
Pillow==8.3.2
polib==1.1.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
python-stdnum==1.17
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tornado==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==1.3.0
wcwidth==0.2.5
zipp==3.6.0
